# src/main/resources/swagger/library-management-v1.yaml
openapi: 3.0.3
info:
  title: Library Management API
  description: API for managing books, members, and book loans
  version: 1.0.0
servers:
  - url: http://localhost:5004
    description: Local development server
...
components:
  schemas:
    BookDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          maxLength: 200
        author:
          type: string
          maxLength: 100
        isbn:
          type: string
          pattern: '^[0-9-]{13,17}$'
        publishYear:
          type: integer
          minimum: 1000
          maximum: 9999
        genreIds:
          type: array
          items:
            type: integer
            format: int64
        addedDateTime:
          type: string
          format: date-time
        updatedDateTime:
          type: string
          format: date-time
      required:
        - title
        - author
        - isbn
...
    PageResponseDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BookDTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        pageNumber:
          type: integer
        pageSize:
          type: integer
    ErrorDTO:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
paths:
  /api/v1/books:
    post:
      tags:
        - Books
      summary: Add a new book
      operationId: addBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookDTO'
      responses:
        '201':
          description: Book added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDTO'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

    get:
      tags:
        - Books
      summary: Get all books with pagination
      operationId: getAllBooks
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseDTO'
  /api/v1/books/{id}:
    get:
      tags:
        - Books
      summary: Get book by ID
      operationId: getBookById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDTO'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

    put:
      tags:
        - Books
      summary: Update book
      operationId: updateBook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookDTO'
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDTO'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

    delete:
      tags:
        - Books
      summary: Remove book
      operationId: removeBook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Book removed successfully
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
...